+++ simple-spring/simple-springjetty-demo/src/main/java/com/mimu/simple/sj/request/UserDataRequest.java
@@ -14,6 +14,6 @@ import lombok.NoArgsConstructor;
 @NoArgsConstructor
 @AllArgsConstructor
 public class UserDataRequest {
-    private int pid;
+    private long pid;
     private int termId;
 }
